backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-genai
anyio
PyJWT
python-multipart
asyncpg


